Migrate various extension methods on IPublishedContent, IPublishedElement and IPublishedProperty to Umbraco.Abstractions
