+++ src/Umbraco.Abstractions/Models/PublishedContent/IndexedArrayItem.cs
