+++ src/Umbraco.Tests/PublishedContent/PublishedContentExtensionTests.cs
@@ -1,6 +1,7 @@
 ﻿using System.Collections.Generic;
 using NUnit.Framework;
 using Umbraco.Web.Composing;
+using Umbraco.Core;
 using Umbraco.Core.Models;
 using Umbraco.Core.Models.PublishedContent;
 using Umbraco.Tests.Testing;
