+++ src/Umbraco.Web/PublishedPropertyExtension.cs
@@ -1,5 +1,4 @@
-﻿using System.Collections.Generic;
-using Umbraco.Core;
+﻿using Umbraco.Core;
 using Umbraco.Web.Composing;
 using Umbraco.Core.Models.PublishedContent;
 
@@ -18,12 +17,7 @@ namespace Umbraco.Web
 
         public static object Value(this IPublishedProperty property, string culture = null, string segment = null, Fallback fallback = default, object defaultValue = default)
         {
-            if (property.HasValue(culture, segment))
-                return property.GetValue(culture, segment);
-
-            return PublishedValueFallback.TryGetValue(property, culture, segment, fallback, defaultValue, out var value)
-                ? value
-                : property.GetValue(culture, segment); // give converter a chance to return it's own vision of "no value"
+            return property.Value(PublishedValueFallback, culture, segment, fallback, defaultValue);
         }
 
         #endregion
@@ -32,33 +26,7 @@ namespace Umbraco.Web
 
         public static T Value<T>(this IPublishedProperty property, string culture = null, string segment = null, Fallback fallback = default, T defaultValue = default)
         {
-            if (property.HasValue(culture, segment))
-            {
-                // we have a value
-                // try to cast or convert it
-                var value =  property.GetValue(culture, segment);
-                if (value is T valueAsT) return valueAsT;
-                var valueConverted = value.TryConvertTo<T>();
-                if (valueConverted) return valueConverted.Result;
-
-                // cannot cast nor convert the value, nothing we can return but 'default'
-                // note: we don't want to fallback in that case - would make little sense
-                return default;
-            }
-
-            // we don't have a value, try fallback
-            if (PublishedValueFallback.TryGetValue(property, culture, segment, fallback, defaultValue, out var fallbackValue))
-                return fallbackValue;
-
-            // we don't have a value - neither direct nor fallback
-            // give a chance to the converter to return something (eg empty enumerable)
-            var noValue = property.GetValue(culture, segment);
-            if (noValue is T noValueAsT) return noValueAsT;
-            var noValueConverted = noValue.TryConvertTo<T>();
-            if (noValueConverted) return noValueConverted.Result;
-
-            // cannot cast noValue nor convert it, nothing we can return but 'default'
-            return default;
+            return property.Value<T>(PublishedValueFallback, culture, segment, fallback, defaultValue);            
         }
 
         #endregion
